11 - Add some authorization rules

+++ Agora.API/Controllers/TransactionsController.cs
@@ -1,10 +1,12 @@
-﻿using Agora.API.DTOs.Transaction;
+﻿using System.Security.Claims;
+using Agora.API.DTOs.Transaction;
 using Agora.API.InputValidation.Interfaces;
 using Agora.API.Orchestrators.Interfaces;
 using Agora.API.QueryParams;
 using Agora.Core.Interfaces;
 using Agora.Core.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Agora.API.Controllers;
@@ -19,24 +21,49 @@ public class TransactionsController(
     : ControllerBase
 {
     private const string TransactionNotFoundMessage = "Transaction not found.";
+    private const string NotInvolvedMessage = "Current user is not involved in the transaction.";
     
+    // An admin has access to all transactions, but normal user have only access to transactions in which it is involved.
+    [Authorize]
     [HttpGet]
-    public async Task<ActionResult<IReadOnlyList<TransactionSummaryDto>>> GetAllTransactions([FromQuery] TransactionQueryParameters queryParameters)
+    public async Task<ActionResult<IReadOnlyList<TransactionSummaryDto>>> GetAllTransactions(
+        [FromQuery] TransactionQueryParameters queryParameters)
     {
-        IReadOnlyList<Transaction> transactions = await repo.GetAllTransactionsAsync(queryParameters);
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (currentUserId != null)
+        {
+            return Unauthorized("User ID not found in claims.");
+        }
+        
+        bool isAdmin = User.IsInRole("Admin");
+        
+        IReadOnlyList<Transaction> transactions = await repo.GetAllTransactionsAsync(queryParameters, isAdmin ? null : currentUserId);
         return Ok(mapper.Map<IReadOnlyList<TransactionSummaryDto>>(transactions));
     }
 
+    // The current user should either be an admin or be involved in the transaction to see it
+    [Authorize]
     [HttpGet("{id:long}")]
     public async Task<ActionResult<TransactionDetailsDto>> GetTransaction([FromRoute] long id)
     {
         Transaction? transaction = await repo.GetTransactionByIdAsync(id);
 
-        return transaction == null 
-            ? NotFound(TransactionNotFoundMessage) 
-            : Ok(mapper.Map<TransactionDetailsDto>(transaction));
+        if (transaction == null)
+        {
+            return NotFound(TransactionNotFoundMessage);
+        }
+        
+        // Ownership validation
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (transaction.BuyerId != currentUserId && transaction.SellerId != currentUserId)
+        {
+            return Unauthorized(NotInvolvedMessage);
+        }
+        
+        return Ok(mapper.Map<TransactionDetailsDto>(transaction));
     }
     
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<TransactionDetailsDto>> CreateTransaction([FromBody] CreateTransactionDto transactionDto)
     {
@@ -75,6 +102,7 @@ public class TransactionsController(
         return BadRequest("Problem creating the transaction.");
     }
 
+    [Authorize]
     [HttpPut("{id:long}")]
     public async Task<ActionResult> UpdateTransaction([FromRoute] long id, [FromBody] UpdateTransactionDto transactionDto)
     {
@@ -83,6 +111,13 @@ public class TransactionsController(
         if (existingTransaction == null)
         {
             return NotFound(TransactionNotFoundMessage);
+        }
+        
+        // Ownership validation
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (existingTransaction.BuyerId != currentUserId && existingTransaction.SellerId != currentUserId)
+        {
+            return Unauthorized(NotInvolvedMessage);
         }
         
         // Input validation
@@ -108,6 +143,7 @@ public class TransactionsController(
             : BadRequest("Problem updating the transaction.");
     }
 
+    [Authorize]
     [HttpDelete("{id:long}")]
     public async Task<ActionResult> DeleteTransaction([FromRoute] long id)
     {
@@ -116,6 +152,13 @@ public class TransactionsController(
         if (transaction == null)
         {
             return NotFound(TransactionNotFoundMessage);
+        }
+        
+        // Ownership validation
+        string? currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (transaction.BuyerId != currentUserId && transaction.SellerId != currentUserId)
+        {
+            return Unauthorized(NotInvolvedMessage);
         }
 
         repo.DeleteTransaction(transaction);
